Move impls for primitives into separate module
